docs(button): combine basic, size button story
