+++ packages/design-system/src/stories/components/Button/BasicButton.stories.tsx
@@ -1,9 +1,23 @@
 import React from "react";
 import { action } from "@storybook/addon-actions";
+import {
+  Table,
+  TableHead,
+  TableRow,
+  TableBody,
+  TableCell,
+  Typography,
+} from "@mui/material";
+import { Bell } from "@lunit/design-system-icons";
 
 import Button from "@/components/Button";
 
 import type { ComponentStory, ComponentMeta } from "@storybook/react";
+import type { ButtonProps } from "@/components/Button/Button.types";
+
+type Size = Pick<ButtonProps, "size">;
+type SizeValues = Size[keyof Size];
+const sizeList: SizeValues[] = ["small", "medium", "large"];
 
 export default {
   title: "Components/Button",
@@ -111,6 +125,48 @@ export default {
   },
 } as ComponentMeta<typeof Button>;
 
-export const BasicButton: ComponentStory<typeof Button> = (arg) => (
-  <Button {...arg}>{arg.children}</Button>
+export const BasicButton: ComponentStory<typeof Button> = (args) => (
+  <Button {...args}>{args.children}</Button>
 );
+
+const SizeButtonTemplate: ComponentStory<typeof Button> = (args) => {
+  return (
+    <Table sx={{ width: 900 }}>
+      <TableHead>
+        <TableRow>
+          <TableCell colSpan={3} sx={{ fontWeight: "bold", fontSize: "16px" }}>
+            <Typography variant="body1_sb">Size</Typography>
+          </TableCell>
+        </TableRow>
+        <TableRow>
+          <TableCell>
+            <Typography variant="body2_m">Small</Typography>
+          </TableCell>
+          <TableCell>
+            <Typography variant="body2_m">Medium</Typography>
+          </TableCell>
+          <TableCell>
+            <Typography variant="body2_m">Large</Typography>
+          </TableCell>
+        </TableRow>
+      </TableHead>
+      <TableBody>
+        <TableRow>
+          {sizeList.map((size) => (
+            <TableCell key={size} sx={{ "& button": { marginRight: "10px" } }}>
+              <Button {...args} hasIconOnly icon={<Bell />} size={size} />
+              <Button {...args} size={size}>
+                {args.children}
+              </Button>
+              <Button {...args} icon={<Bell />} size={size}>
+                {args.children}
+              </Button>
+            </TableCell>
+          ))}
+        </TableRow>
+      </TableBody>
+    </Table>
+  );
+};
+
+export const Size = SizeButtonTemplate.bind({});
